requirements.txt
pandas
numpy
sklearn
plotly.express
base64
os
datetime
matplotlib.pyplot
seaborn
pickle
asyncio.windows_events